# swagger.yaml
openapi: 3.0.3
info:
  title: Swagger API Course FE
  description: |-
    This is a list api of course FE
  contact:
    email: lttd0204@gmail.com
  version: 1.0.11
servers:
  - url: https://api-shop-l5ir.onrender.com
...
paths:
...
  /api/comments/{commentId}:
    get:
      tags:
        - comments
      summary: Get details comment by ID
      description: Get details comment by id of comments
      operationId: getCommentById
      parameters:
        - name: commentId
          in: path
          description: ID of comment that needs to be fetched
          required: true
          schema:
            type: string
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
            application/xml:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
        "401":
          description: UNAUTHORIZED
        "500":
          description: INTERNAL_SERVER_ERROR
        "409":
          description: ALREADY_EXIST
        "400":
          description: INVALID
      security:
        - bearerAuth: []
    delete:
      tags:
        - comments
      summary: Delete comment by ID
      description: Try delete comment by id of comment
      operationId: deleteComment
      parameters:
        - name: commentId
          in: path
          description: ID of the comment that needs to be deleted
          required: true
          schema:
            type: string
      responses:
        "201":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
            application/xml:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
        "401":
          description: UNAUTHORIZED
        "500":
          description: INTERNAL_SERVER_ERROR
        "409":
          description: ALREADY_EXIST
        "400":
          description: INVALID
      security:
        - bearerAuth: []
    put:
      tags:
        - comments
      summary: Update an existing comment
      description: Update an existing comment by Id
      operationId: updateComment
      parameters:
        - name: commentId
          in: path
          description: comment id of comment updated
          required: true
          schema:
            type: string
      requestBody:
        description: Update an existent comment in the store
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Comment"
          application/xml:
            schema:
              $ref: "#/components/schemas/Comment"
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/Comment"
        required: true
      responses:
        "201":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
            application/xml:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
        "401":
          description: UNAUTHORIZED
        "500":
          description: INTERNAL_SERVER_ERROR
        "409":
          description: ALREADY_EXIST
        "400":
          description: INVALID
      security:
        - bearerAuth: []
  /api/comment/me/{commentId}:
    put:
      tags:
        - comments
      summary: Update an existing my comment
      description: Update an existing my comment
      operationId: updateComment
      parameters:
        - name: commentId
          in: path
          description: comment id of comment updated
          required: true
          schema:
            type: string
      requestBody:
        description: Update an existent comment in the store
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Comment"
          application/xml:
            schema:
              $ref: "#/components/schemas/Comment"
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/Comment"
        required: true
      responses:
        "201":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
            application/xml:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
        "401":
          description: UNAUTHORIZED
        "500":
          description: INTERNAL_SERVER_ERROR
        "409":
          description: ALREADY_EXIST
        "400":
          description: INVALID
      security:
        - bearerAuth: []

    delete:
      tags:
        - comments
      summary: Delete my comment
      description: Try delete my comment by id of you
      operationId: deleteComment
      parameters:
        - name: commentId
          in: path
          description: ID of the comment that needs to be deleted
          required: true
          schema:
            type: string
      responses:
        "201":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
            application/xml:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
        "401":
          description: UNAUTHORIZED
        "500":
          description: INTERNAL_SERVER_ERROR
        "409":
          description: ALREADY_EXIST
        "400":
          description: INVALID
      security:
        - bearerAuth: []
...
  /api/notifications/{notificationId}:
    delete:
      tags:
        - notifications
      summary: Delete notification by id
      description: Delete notification by id
      operationId: deleteNotification
      parameters:
        - name: productId
          in: path
          description: ID of notification that needs to be fetched
          required: true
          schema:
            type: string
      responses:
        "201":
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: "#/components/schemas/Notification"
            application/xml:
              schema:
                type: object
                items:
                  $ref: "#/components/schemas/Notification"
        "401":
          description: UNAUTHORIZED
        "500":
          description: INTERNAL_SERVER_ERROR
        "409":
          description: ALREADY_EXIST
        "400":
          description: INVALID
      security:
        - bearerAuth: []
...
components:
  schemas:
...
    Notification:
      type: object
      properties:
        _id:
          type: string
          format: uuid
        recipientIds:
          type: array
          items:
            type: object
            properties:
              userId:
                type: string
                format: uuid
              isRead:
                type: boolean
        referenceId:
          type: string
          format: uuid
        body:
          type: string
        title:
          type: string
        context:
          type: string
      required:
        - context
        - title
        - body
      xml:
        name: Notification
...
    Comment:
      type: object
      properties:
        content:
          type: string
        product:
          type: string
          format: uuid
        user:
          type: string
          format: uuid
      required:
        - content
        - user
        - product
      xml:
        name: Review
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT